Add Debug3dRenderer
Allows to see Box or Camera in 3d space.
This patch remove BoxHelper and CameraHelper because
Debug3dRenderer replace its.
